helm/util/requirements.txt
# Tooling for keeping the helm-chart up-to-date
semver==3.0.1
frigate==0.7.0
toml==0.10.2
pyyaml==6.0.1
